feat: add reset function for load photos
