+++ js/form-send.js
@@ -3,7 +3,8 @@ import { openSuccessSendMessage, openErrorSendMessage } from './form-messages.js
 import { sendData } from './api.js';
 import { resetMainPin, mainPinLocation } from './map.js';
 import { getLocationToString } from './utils.js';
-import { mainPoint, FILE_TYPES } from './const.js';
+import { mainPoint, NUMBER_AFTER_POINT } from './const.js';
+import { clearAllLoadPhoto } from './load-images.js';
 
 const SubmitButtonText = {
   IDLE: 'Сохранить',
@@ -13,16 +14,12 @@ const SubmitButtonText = {
 const resetFormButton = document.querySelector('.ad-form__reset');
 const submitButton = document.querySelector('.ad-form__submit');
 
-const avatarChooser = document.querySelector('.ad-form__field input[type=file]');
-const avatarPreview = document.querySelector('.ad-form-header__preview img');
-const photoChooser = document.querySelector('.ad-form__upload input[type=file]');
-const photoPreview = document.querySelector('.ad-form__photo');
-
 const resetForm = (evt) => {
   evt.preventDefault();
   pristine.reset();
+  clearAllLoadPhoto();
   adForm.reset();
-  mainPinLocation.value = getLocationToString(mainPoint, 5);
+  mainPinLocation.value = getLocationToString(mainPoint, NUMBER_AFTER_POINT);
   resetMainPin();
 };
 
@@ -61,27 +58,5 @@ const setUserFormSubmit = () => {
     }
   });
 };
-
-avatarChooser.addEventListener('change', () => {
-  const avatar = avatarChooser.files[0];
-  const fileName = avatar.name.toLowerCase();
-
-  const matches = FILE_TYPES.some((it) => fileName.endsWith(it));
-
-  if (matches) {
-    avatarPreview.src = URL.createObjectURL(avatar);
-  }
-});
-
-photoChooser.addEventListener('change', () => {
-  const photo = photoChooser.files[0];
-  const photoName = photo.name.toLowerCase();
-
-  const matchTypes = FILE_TYPES.some((it) => photoName.endsWith(it));
-
-  if (matchTypes) {
-    photoPreview.innerHTML = `<img src="${URL.createObjectURL(photo)}" width="70" height="70">`;
-  }
-});
 
 export { setUserFormSubmit };
